# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.ProductInput:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - stock
    type: object
  handlers.RegisterInput:
    properties:
      password:
        type: string
      roles:
        items:
          type: string
        minItems: 1
        type: array
      username:
        type: string
    required:
    - password
    - roles
    - username
    type: object
  handlers.StoreInput:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  handlers.SubmitReviewInput:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      reviewer_name:
        type: string
    required:
    - comment
    - rating
    - reviewer_name
    type: object
  models.AppReview:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      reviewer_name:
        type: string
    type: object
  models.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      store_id:
        type: integer
    type: object
  models.Store:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Interactive API documentation for the SEAPEDIA marketplace system.
  title: SEAPEDIA API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in user and return JWT token alongside registered roles
      parameters:
      - description: Login Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Returns JWT Token & Role list
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid credentials schema
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/profile:
    get:
      description: Fetches current credentials using authentication token
      produces:
      - application/json
      responses:
        "200":
          description: User profile details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get logged-in user profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user profile with selected roles (Admin, Seller,
        Buyer, Driver)
      parameters:
      - description: Registration Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role or request data
          schema:
            additionalProperties: true
            type: object
        "499":
          description: Username already exists
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
  /catalog:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: View public catalog
      tags:
      - Public Catalog
  /catalog/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: View product details
      tags:
      - Public Catalog
  /reviews:
    get:
      description: Get a list of all public application and website feedback
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews
          schema:
            items:
              $ref: '#/definitions/models.AppReview'
            type: array
      summary: List application reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Submit feedback about the website or application experience. Rating
        must be between 1 and 5.
      parameters:
      - description: Review Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SubmitReviewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Review submitted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Submit an application review
      tags:
      - Reviews
  /seller/products:
    get:
      parameters:
      - description: Active user role should be Buyer
        in: header
        name: X-Active-Role
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      security:
      - BearerAuth: []
      summary: List seller products
      tags:
      - Store Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerAuth: []
      summary: Add a new product
      tags:
      - Store Products
  /seller/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Store Products
    put:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ProductInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerAuth: []
      summary: Update product details
      tags:
      - Store Products
  /seller/store:
    post:
      consumes:
      - application/json
      description: Register a store with a unique name. Active role must be 'Seller'.
      parameters:
      - description: Store Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.StoreInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Store'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Store name already exists or user already owns a store
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a Store profile
      tags:
      - Store
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"